resistance calculations and experimentation, igniter test - worked, external buttons and button class test - works, fixing some bugs with analog read

+++ electronics/static_fire/lib/Devices/LoadCell/LoadCell.cpp
@@ -51,7 +51,7 @@ float LoadCell::getAveragedSamples(uint16_t numReadings) {
 }
 
 float LoadCell::getSample_mV() {
-  float newReading = analogReadMilliVolts(m_sensorPin);  // !!! test this
+  float newReading = (float)analogReadMilliVolts(m_sensorPin);  // !!! test this
   // any corrections
   return newReading;
 }
